requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-social-auth==0.2.12
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
SHop==0.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
